Added transport and resource_id modules
