+++ src/lib.rs
@@ -1,6 +1,7 @@
 mod network_adapter;
 mod encoding;
 mod adapter;
-mod tcp;
+mod transports;
+mod resource_id;
 pub mod network;
 pub mod events;
